+++ Scripts/Data Classes/Transaction.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 
 public class Transaction
 {
-    public int transactionID;
+    public long transactionID;
     public DateTime transactionDateTime;
     public float transactionAmount;
     public int paymentType;
@@ -14,7 +14,7 @@ public class Transaction
     public List<(Item, int)> transactionItems = new List<(Item, int)>(); //item and quantity (place price in item)
 
     //Initilsing procedure
-    public Transaction(int transIDToAdd, DateTime transDateTimeToAdd, float transAmountToAdd, int payTypeToAdd, int staffIDToAdd, List<(Item, int)> itemsToAdd)
+    public Transaction(long transIDToAdd, DateTime transDateTimeToAdd, float transAmountToAdd, int payTypeToAdd, int staffIDToAdd, List<(Item, int)> itemsToAdd)
     {
         transactionID = transIDToAdd;
         transactionDateTime = transDateTimeToAdd;
